# swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "API de TO-DO LIST - Welinton Suarez - Diana Bedoya",
    "version": "1.0.0",
    "description": "Una API para administrar un To-Do list"
  },
  "servers": [
    {
      "url": "http://localhost:3001",
      "description": "Servidor de desarrollo"
    }
  ],
  "paths": {
    "/": {
      "get": {
        "summary": "Página de inicio",
        "description": "Retorna un mensaje de bienvenida",
        "responses": {
          "200": {
            "description": "Respuesta exitosa",
            "content": {
              "text/html": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
    "/tasks": {
      "get": {
        "summary": "Obtener todas la tareas",
        "description": "Retorna una lista de todos las tareas",
        "responses": {
          "200": {
            "description": "Respuesta exitosa",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Task"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Crear una nueva tarea",
        "description": "Crea una nueva tarea con la información proporcionada",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TaskInput"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Tarea creada exitosamente",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Task"
                }
              }
            }
          }
        }
      }
    },
    "/tasks/{id}": {
      "get": {
        "summary": "Obtener una tarea por ID",
        "description": "Retorna una tarea en específico basado en su ID",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "schema": {
              "type": "string"
            },
            "required": true,
            "description": "ID de la tarea"
          }
        ],
        "responses": {
          "200": {
            "description": "Respuesta exitosa",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Task"
                }
              }
            }
          },
          "404": {
            "description": "Tarea no encontrada",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "description": "Mensaje de error"
                    }
                  }
                }
              }
            }
          }
        }
      },
      "put": {
        "summary": "Actualizar una tarea existente",
        "description": "Actualiza la información de una tarea existente basado en su ID",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "schema": {
              "type": "string"
            },
            "required": true,
            "description": "ID del contacto"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TaskInput"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Tarea actualizada exitosamente",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Task"
                }
              }
            }
          },
          "404": {
            "description": "Tarea no encontrada",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "description": "Mensaje de error"
                    }
                  }
                }
              }
            }
          }
        }
      },
      "delete": {
        "summary": "Elimina una tarea correctamente",
        "description": "Elimina una tarea existente basado en su ID",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "schema": {
              "type": "string"
            },
            "required": true,
            "description": "ID del contacto"
          }
        ],
        "responses": {
          "204": {
            "description": "Tarea eliminada exitosamente"
          },
          "404": {
            "description": "Tarea no encontrada",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "description": "Mensaje de error"
                    }
                  }
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "Task": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "ID de la tarea"
          },
          "task": {
            "type": "string",
            "description": "Descripcion de la tarea"
          },
          "assigned": {
            "type": "string",
            "description": "Nombre de la persona a cargo de la tarea"
          },
          "state": {
            "type": "string",
            "description": "Estado de la tarea (to-do, in-progress , done)"
          }
        }
      },
      "TaskInput": {
        "type": "object",
        "properties": {
          "task": {
            "type": "string",
            "description": "Descripcion de la tarea"
          },
          "assigned": {
            "type": "string",
            "description": "Nombre de la persona a cargo de la tarea"
          },
          "state": {
            "type": "string",
            "description": "Estado de la tarea (to-do, in-progress , done)"
          }
        },
        "required": [
          "title",
          "status",
          "responsible"
        ]
      }
    }
  },
  "tags": []
}